% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_osf.R
\name{build_osf}
\alias{build_osf}
\title{Build a new OSF project}
\usage{
build_osf(title, tokenid, tokentype = "file")
}
\arguments{
\item{title}{name of project to build on OSF}

\item{tokenid}{name associated with token file stored on Server (otherwise your token password)}

\item{tokentype}{file or password (default = file)}
}
\description{
This function will build a new OSF project
}
\examples{

}
\keyword{osf}
