% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files_to_osf.R
\name{check_files_to_osf}
\alias{check_files_to_osf}
\title{Check which files will be uploaded to OSF}
\usage{
check_files_to_osf(dir = ".", skip_files = c(), recursive = TRUE)
}
\arguments{
\item{dir}{A file path to a first-level directory in your local project
    folder. e.g., "data". Do not specifiy second-level directories
    e.g., "data/raw".

    To upload files from the root project folder, specify dir = "." or leave
    out argument.}

\item{skip_files}{An array of file/folder names and pattern strings to not
include in the upload to OSF. e.g., you might have a folder of "old"
scripts or data files that you do not want uploaded.}

\item{recursive}{Search files recursively? Set to FALSE if you only want the
files in a first-level directory and not any further sub-directories.
e.g., "data" or ".".}
}
\description{
Verify that the file/folder names and pattern strings specificed in
upload_to_osf(skip_files = c()) argument will only upload the files you want
to upload.
}
