% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{copy_files}
\alias{copy_files}
\title{Copy data files from one direcoty to another}
\usage{
copy_files(
  from,
  to,
  filetype,
  sub_folder = NULL,
  task_dir.names = c("none", "numbered", "asis"),
  remove = FALSE,
  copy = TRUE,
  overwrite = c("no", "yes", "skip"),
  ignore = NULL,
  recursive = TRUE
)
}
\arguments{
\item{from}{a directory of task folders}

\item{to}{a directory where raw data files are to be copied to}

\item{filetype}{a string pattern identifying a specific
type of raw data file (e.g. .edat2).}

\item{sub_folder}{optional. do you want to copy files to a sub folder within
the `to = ` directory?}

\item{task_dir.names}{Relationship between task name and task folder name.
Default is "none". Alternative options are; "numbered", "asis".}

\item{remove}{logical. Should the files in the 'from' directory
be deleted? (Default: FALSE)}

\item{copy}{logical. Do you want to copy the files? (Default = TRUE)}

\item{overwrite}{"no" (default), "yes", or "skip". "no" will copy over the
file but append the file with "DUPLICATED". "yes" will overwrite the
file. "skip" will not copy the file over.}

\item{ignore}{a character vector of task folders to ignore}

\item{recursive}{logical. Should the function search for files recursively?}
}
\description{
This function will automatically copy files from the Task directory to
the Raw data directory
}
