% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Setup directory and templates for a new research study}
\usage{
create_project(
  type = "data analysis",
  sessions = 1,
  figures.dir = FALSE,
  manuscript.dir = FALSE,
  presentations.dir = FALSE,
  other.dir = c(),
  mainscript = NULL,
  rawscript = FALSE,
  scorescript = FALSE,
  mergescript = FALSE,
  path = "."
)
}
\arguments{
\item{type}{String. Options: "data collection"; "data analysis".}

\item{sessions}{Numeric. How many sessions will the study have?}

\item{figures.dir}{Logical. Create figures directory? default = FALSE}

\item{manuscript.dir}{Logical. Create manuscript directory? default = FALSE}

\item{presentations.dir}{Logical. Create presentations directory? default = FALSE}

\item{other.dir}{List of other directories you want to create}

\item{mainscript}{Logical. Download mainscript template? default = TRUE}

\item{rawscript}{Logical. Download rawscript template? default = FALSE}

\item{scorescript}{Logical. Download scorescript template? default = FALSE}

\item{mergescript}{Logical. Download mergescript template? default = FALSE}

\item{path}{String. Home directory file path}

\item{scripts.dir}{Logical. Create script directory? default = TRUE}

\item{data.dir}{Logical. Create data files directory? default = TRUE}

\item{raw.dir}{Logical. Create raw? default = TRUE}

\item{messy.dir}{Logical. Create raw messy directory? default = TRUE}

\item{messy.name}{String. Name of messy direcotry. default = "messy"}

\item{scored.dir}{Logical. Create scored directory? default = TRUE}

\item{tasks.dir}{Logical. Create tasks directory? default = TRUE}

\item{documents.dir}{Logical. Create documents directory? default = TRUE}

\item{results.dir}{Logical. Create results directory? default = TRUE}

\item{generic}{Logical. Download all generic templates? default = FALSE}
}
\description{
This function can be used to automatically setup your study directory by
creating folders and template scripts
}
