% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_osf_project.R, R/create_study.R,
%   R/upload_file_osf.R
\name{create_study}
\alias{create_study}
\title{Create Directory Organization For a New Study test}
\usage{
create_study()

create_study()

create_study()
}
\arguments{
\item{scripts.dir}{Logical. Create script directory? default = TRUE}

\item{data.dir}{Logical. Create data files directory? default = TRUE}

\item{raw.dir}{Logical. Create raw? default = TRUE}

\item{messy.dir}{Logical. Create raw messy directory? default = TRUE}

\item{messy.name}{String. Name of messy direcotry. default = "messy"}

\item{scored.dir}{Logical. Create scored directory? default = TRUE}

\item{tasks.dir}{Logical. Create tasks directory? default = TRUE}

\item{documents.dir}{Logical. Create documents directory? default = TRUE}

\item{results.dir}{Logical. Create results directory? default = TRUE}

\item{figures.dir}{Logical. Create figures directory? default = TRUE}

\item{manuscript.dir}{Logical. Create manuscript directory? default = TRUE}

\item{presentations.dir}{Logical. Create presentations directory? default = TRUE}

\item{other.dir}{List of other directories you want to create}

\item{masterscript}{Logical. Download masterscript template? default = TRUE}

\item{rawscript}{Logical. Download rawscript template? default = TRUE}

\item{scorescript}{Logical. Download scorescript template? default = TRUE}

\item{mergescript}{Logical. Download mergescript template? default = TRUE}

\item{demographics}{Logical. Download demographics template? default = TRUE}

\item{sem}{Logical. Download sem templates? default = TRUE}

\item{scripts.dir}{Logical. Create script directory? default = TRUE}

\item{data.dir}{Logical. Create data files directory? default = TRUE}

\item{raw.dir}{Logical. Create raw? default = TRUE}

\item{messy.dir}{Logical. Create raw messy directory? default = TRUE}

\item{messy.name}{String. Name of messy direcotry. default = "messy"}

\item{scored.dir}{Logical. Create scored directory? default = TRUE}

\item{tasks.dir}{Logical. Create tasks directory? default = TRUE}

\item{documents.dir}{Logical. Create documents directory? default = TRUE}

\item{results.dir}{Logical. Create results directory? default = TRUE}

\item{figures.dir}{Logical. Create figures directory? default = TRUE}

\item{manuscript.dir}{Logical. Create manuscript directory? default = TRUE}

\item{presentations.dir}{Logical. Create presentations directory? default = TRUE}

\item{other.dir}{List of other directories you want to create}

\item{masterscript}{Logical. Download masterscript template? default = TRUE}

\item{rawscript}{Logical. Download rawscript template? default = TRUE}

\item{scorescript}{Logical. Download scorescript template? default = TRUE}

\item{mergescript}{Logical. Download mergescript template? default = TRUE}

\item{demographics}{Logical. Download demographics template? default = TRUE}

\item{sem}{Logical. Download sem templates? default = TRUE}

\item{scripts.dir}{Logical. Create script directory? default = TRUE}

\item{data.dir}{Logical. Create data files directory? default = TRUE}

\item{raw.dir}{Logical. Create raw? default = TRUE}

\item{messy.dir}{Logical. Create raw messy directory? default = TRUE}

\item{messy.name}{String. Name of messy direcotry. default = "messy"}

\item{scored.dir}{Logical. Create scored directory? default = TRUE}

\item{tasks.dir}{Logical. Create tasks directory? default = TRUE}

\item{documents.dir}{Logical. Create documents directory? default = TRUE}

\item{results.dir}{Logical. Create results directory? default = TRUE}

\item{figures.dir}{Logical. Create figures directory? default = TRUE}

\item{manuscript.dir}{Logical. Create manuscript directory? default = TRUE}

\item{presentations.dir}{Logical. Create presentations directory? default = TRUE}

\item{other.dir}{List of other directories you want to create}

\item{masterscript}{Logical. Download masterscript template? default = TRUE}

\item{rawscript}{Logical. Download rawscript template? default = TRUE}

\item{scorescript}{Logical. Download scorescript template? default = TRUE}

\item{mergescript}{Logical. Download mergescript template? default = TRUE}

\item{demographics}{Logical. Download demographics template? default = TRUE}

\item{sem}{Logical. Download sem templates? default = TRUE}
}
\description{
This function can be used to automatically setup your study directory by
creating folders and template scripts

This function can be used to automatically setup your study directory by
creating folders and template scripts

This function can be used to automatically setup your study directory by
creating folders and template scripts
}
\examples{



}
\keyword{center}
\keyword{create}
