% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_check.R
\name{duplicates_check}
\alias{duplicates_check}
\title{Check and Remove Duplicate IDs}
\usage{
duplicates_check(
  x,
  unique_id = "Subject",
  date_time = c("SessionDate", "SessionTime"),
  n = 1,
  remove = TRUE,
  keep_by = c("none", "first date", "least missing"),
  save_as = NULL,
  id = "Subject",
  unique = c("SessionDate", "SessionTime")
)
}
\arguments{
\item{x}{dataframe}

\item{unique_id}{Column name(s) that should be uniquely identified and will be used to check for
duplicate id's grouped by these columns (e.g., "Subject", c("Subject, Session))}

\item{date_time}{Column names that uniquely identify a testing administration.
(e.g., "Date", "Time"). Can also be: "none" or NULL}

\item{n}{Number of unique id's expected (default: 1).}

\item{remove}{logical. Remove duplicate ids from data? (default: TRUE)}

\item{keep_by}{Which duplicate id should be kept?
options: "none", "first date", "least missing"}

\item{save_as}{Folder path and file name to output the duplicate ID's}

\item{id}{deprecated. Use unique_id instead.}

\item{unique}{deprecated. Use date_time instead.}
}
\description{
Checks and removes duplicate subject ids that can happen when misentered
during task administration
}
