% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_bind.R
\name{files_bind}
\alias{files_bind}
\title{Bind Multiple Files}
\usage{
files_bind(
  path = "",
  pattern = "",
  delim = ",",
  bind = "rows",
  output_file = "",
  source = ""
)
}
\arguments{
\item{path}{Folder location of files to be binded}

\item{pattern}{Pattern to identify files to be binded}

\item{delim}{Delimiter used in files. Passed onto `readr::read_delim()`}

\item{bind}{Bind "rows" or "columns"?}

\item{output_file}{File name and path to be saved to.}

\item{source}{If data file comes from a specific software specify it here
for better support. Currently supports "eprime".}
}
\description{
Binds multiple files in the same directory into a single data frame.
Bind can occur by "rows" or "columns".
}
