% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Download R Script Templates}
\usage{
get_template(
  to = "R Scripts/templates",
  overwrite = FALSE,
  mainscript = NULL,
  rawscript = FALSE,
  scorescript = FALSE,
  mergescript = FALSE,
  path = "."
)
}
\arguments{
\item{to}{a directory where to download R scripts.
default: "R Scripts/templates"}

\item{overwrite}{Logical. Overwite any existing templates? default: FALSE}

\item{mainscript}{String. Which type of mainscript template to download.
options: "data preparation" or "data analysis".}

\item{rawscript}{Logical. Download template to convert a "messy" raw data file
to a "tidy" raw data file? default = FALSE}

\item{scorescript}{Logical. Download template to perform data cleaning and
scoring (aggregate) on a "tidy" raw data file? default = FALSE}

\item{mergescript}{Logical. Download template to merge multiple scored data
files? default = FALSE}

\item{path}{String. working directory file path. Ignore...}
}
\description{
This function will download template scripts for data analysis in R
}
