% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_to_visualangle.R
\name{pixel_to_visualangle}
\alias{pixel_to_visualangle}
\title{Check which files will be uploaded to OSF}
\usage{
pixel_to_visualangle(
  pixels,
  monitor_res,
  monitor_diagonal_in,
  viewing_distance
)
}
\arguments{
\item{pixels}{Pixel value to convert to visual angle.}

\item{monitor_res}{A vector of two integers representing the monitor
resolution in pixels. e.g., c(1920, 1080).}

\item{monitor_diagonal_in}{The diagonal size of the monitor in inches.}

\item{viewing_distance}{The distance from the monitor to the viewer in cm.}
}
\description{
Verify that the file/folder names and pattern strings specificed in
upload_to_osf(skip_files = c()) argument will only upload the files you want
to upload.
}
