% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_osf.R
\name{sync_osf}
\alias{sync_osf}
\title{Sync to an OSF project}
\usage{
sync_osf(components, tokenid, tokentype = "file")
}
\arguments{
\item{tokenid}{name associated with token file stored on Server (otherwise your token password)}

\item{tokentype}{file or password (default = file)}
}
\description{
This function will sync to an existing OSF project
}
\examples{

}
\keyword{osf}
