% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Download template R scripts}
\usage{
template(to = "R Scripts/templates", overwrite = FALSE,
  masterscript = NULL, rawscript = FALSE, scorescript = FALSE,
  mergescript = FALSE, path = "./")
}
\arguments{
\item{to}{a directory where to download R scripts}

\item{overwrite}{Logical. Overwite any existing templates?}

\item{masterscript}{String. Which type of masterscript template to download.
masterscript = "data preparation" or masterscript = "data analysis".}

\item{rawscript}{Logical. Download rawscript template? default = FALSE}

\item{scorescript}{Logical. Download scorescript template? default = FALSE}

\item{mergescript}{Logical. Download mergescript template? default = FALSE}

\item{path}{String. Home directory file path. Ignore...}
}
\description{
This function will download R script templates
}
