% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_raw.R
\name{copy_raw}
\alias{copy_raw}
\title{Copy raw data files from one direcoty to another}
\usage{
copy_raw(from, to, pattern, subj.list = NULL, remove = FALSE)
}
\arguments{
\item{from}{a directory of task folders}

\item{to}{a directory where raw data files are to be copied to}

\item{pattern}{a string pattern identifying a specific
type of raw data file (e.g. .edat2).}

\item{subj.list}{a file path to a list of subjects who completed
the entire study. (Default: NULL).}

\item{remove}{logical. Should the files in the 'from' directory
be deleted? (Default: FALSE)}
}
\description{
This function will automatically copy files from the Task directory to
the Raw data directory
}
