% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_raw.R
\name{copy_raw}
\alias{copy_raw}
\title{Copy raw data files from one direcoty to another}
\usage{
copy_raw(from, to, filetype, remove = FALSE, subj.file = NULL,
  copy = TRUE, ext = NULL, task_dir.names = "numbered")
}
\arguments{
\item{from}{a directory of task folders}

\item{to}{a directory where raw data files are to be copied to}

\item{filetype}{a string pattern identifying a specific
type of raw data file (e.g. .edat2).}

\item{remove}{logical. Should the files in the 'from' directory
be deleted? (Default: FALSE)}

\item{subj.file}{a file path to a list of subjects who completed
the entire study. (Default: NULL).}

\item{copy}{logical. Do you want to copy or only remove the files? (Default = TRUE)}

\item{ext}{a directory extension to the "to" argument.}

\item{task_dir.names}{Relationship between task name and task folder name.
Default is "numbered". Alternative options are; "asis".}
}
\description{
This function will automatically copy files from the Task directory to
the Raw data directory
}
