% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Setup directory and templates for a new research study}
\usage{
create_project(
  script_templates = FALSE,
  mainscript = FALSE,
  rawscript = FALSE,
  scorescript = FALSE,
  mergescript = FALSE,
  analysisscript = FALSE,
  study_doc = FALSE,
  data_raw = FALSE,
  data_scored = FALSE,
  documents = FALSE,
  results = FALSE,
  tasks = FALSE,
  sessions = NULL,
  manuscript = FALSE,
  presentations = FALSE,
  figures = FALSE,
  other = c(),
  standard_project = FALSE,
  path = "."
)
}
\arguments{
\item{script_templates}{Logical. Download all script templates? default = FALSE}

\item{mainscript}{Logical. Download a main script template? default = TRUE}

\item{rawscript}{Logical. Download a raw script template? default = FALSE}

\item{scorescript}{Logical. Download a score script template? default = FALSE}

\item{mergescript}{Logical. Download a merge script template? default = FALSE}

\item{analysisscript}{Logical. Download template to do data analysis (.Rmd)? default = FALSE}

\item{study_doc}{Logical. Download a study documentation template? default = FALSE}

\item{data_raw}{Logical. Create a raw data directory? default = FALSE}

\item{data_scored}{Logical. Create a scored data directory? default = FALSE}

\item{documents}{Logical. Create documents directory? default = FALSE}

\item{results}{Logical. Create results directory? default = FALSE}

\item{tasks}{Logical. Create tasks directory? default = FALSE}

\item{sessions}{Numeric. How many sessions will the study have?}

\item{manuscript}{Logical. Create manuscript directory? default = FALSE}

\item{presentations}{Logical. Create presentations directory? default = FALSE}

\item{figures}{Logical. Create figures directory? default = FALSE}

\item{other}{List of other directories you want to create}

\item{standard_project}{Logical. Create a standard project directory? default = FALSE}

\item{path}{String. Home directory file path}
}
\description{
This function can be used to automatically setup your study directory by
creating folders and template scripts
}
