% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study.R
\name{create_study}
\alias{create_study}
\title{Setup directory and templates for a new research study}
\usage{
create_study(type = "data analysis", sessions = 1,
  figures.dir = FALSE, manuscript.dir = FALSE,
  presentations.dir = FALSE, other.dir = c(), masterscript = NULL,
  rawscript = FALSE, scorescript = FALSE, mergescript = FALSE,
  path = ".")
}
\arguments{
\item{type}{String. Options: "data collection"; "analysis".}

\item{sessions}{Numeric. How many sessions will the study have?}

\item{figures.dir}{Logical. Create figures directory? default = FALSE}

\item{manuscript.dir}{Logical. Create manuscript directory? default = FALSE}

\item{presentations.dir}{Logical. Create presentations directory? default = FALSE}

\item{other.dir}{List of other directories you want to create}

\item{masterscript}{Logical. Download masterscript template? default = TRUE}

\item{rawscript}{Logical. Download rawscript template? default = FALSE}

\item{scorescript}{Logical. Download scorescript template? default = FALSE}

\item{mergescript}{Logical. Download mergescript template? default = FALSE}

\item{path}{String. Home directory file path}

\item{scripts.dir}{Logical. Create script directory? default = TRUE}

\item{data.dir}{Logical. Create data files directory? default = TRUE}

\item{raw.dir}{Logical. Create raw? default = TRUE}

\item{messy.dir}{Logical. Create raw messy directory? default = TRUE}

\item{messy.name}{String. Name of messy direcotry. default = "messy"}

\item{scored.dir}{Logical. Create scored directory? default = TRUE}

\item{tasks.dir}{Logical. Create tasks directory? default = TRUE}

\item{documents.dir}{Logical. Create documents directory? default = TRUE}

\item{results.dir}{Logical. Create results directory? default = TRUE}

\item{generic}{Logical. Download all generic templates? default = FALSE}

\item{demographics}{Logical. Download demographics template? default = FALSE}

\item{gf}{Logical. Do you want to download gf template files?}

\item{wmc}{Logical. Do you want to download wmc template files?}

\item{ac}{Logical. Do you want to download all attention control template files?}

\item{antisaccade}{Logical. Do you want to download antisaccade template files?}

\item{stroop}{Logical. Do you want to download stroop template files?}

\item{flanker}{Logical. Do you want to download flanker template files?}

\item{stroopDL}{Logical. Do you want to download stroopDL template files?}

\item{flankerDL}{Logical. Do you want to download flankerDL template files?}

\item{va4}{Logical. Do you want to download va4 template files?}

\item{sact}{Logical. Do you want to download sact template files?}
}
\description{
This function can be used to automatically setup your study directory by
creating folders and template scripts
}
